chore(refactor): extract /api/sessions/* into routes/sessions.ts + http-utils
First bucket of the index.ts route extraction series from the audit.

server/src/http-utils.ts (new) — sendJson / sendError / readJsonBody /
rejectIfNonLocalOrigin + HttpError, factored out of index.ts. The helpers
are returned by makeRouteCtx(req, res) so route modules can destructure
what they need without rebuilding the closures themselves.

server/src/routes/sessions.ts (new) — the seven /api/sessions/* handlers:
  GET /api/sessions
  GET /api/sessions/search
  GET /api/sessions/:id
  GET /api/sessions/:id/events
  GET /api/sessions/:id/events/:eventId
  GET /api/sessions/:id/artifacts
  GET /api/sessions/:id/memory
Same handler bodies; tryHandleSessionRoute returns true when matched, so
the index falls through to other routes when nothing here applies.

Net: index.ts -332 lines (15% smaller). +73 total LOC across the three
files due to per-module ceremony — the helpers will be paid back from the
next bucket onward (artifacts, spaces, memories, auth) since each can
import from http-utils instead of re-inlining the same boilerplate.

No behavioural change. /api/artifacts/:id/sessions stays in index.ts —
it lives under the artifacts URL prefix and will move with that bucket.
