+++ server/src/index.ts
@@ -25,6 +25,8 @@ import { SpaceService } from "./space-service.js";
 import { slugify } from "./utils.js";
 import { IconGenerator } from "./icon-generator.js";
 import { injectBridge } from "./error-bridge.js";
+import { HttpError, makeRouteCtx } from "./http-utils.js";
+import { tryHandleSessionRoute } from "./routes/sessions.js";
 import type { UiCommand } from "../../shared/types.js";
 import {
   scanExistingArtifacts,
@@ -649,72 +651,17 @@ async function handleHttpRequest(req: IncomingMessage, res: ServerResponse) {
 
   const url = req.url || "/";
 
-  // ── Shared helpers ──
-  // Defined near the top so early routes (e.g. GET /api/artifacts) can use
-  // them too, not just the later mutation routes.
+  // Per-request response helpers (sendJson / sendError / readJsonBody /
+  // rejectIfNonLocalOrigin). Live in http-utils.ts so route modules can
+  // share them; HttpError is also exported there.
+  const ctx = makeRouteCtx(req, res);
+  const { sendJson, sendError, readJsonBody, rejectIfNonLocalOrigin } = ctx;
 
-  const sendJson = (data: unknown, status = 200) => {
-    res.writeHead(status, { "Content-Type": "application/json" });
-    res.end(JSON.stringify(data));
-  };
-
-  // Throwable that carries an HTTP status — lets readJsonBody and others
-  // surface a specific status (e.g. 413) without writing the response
-  // themselves, which would race the caller's own catch-block response.
-  class HttpError extends Error {
-    constructor(message: string, public status: number) { super(message); this.name = "HttpError"; }
-  }
-  const sendError = (err: unknown, fallback = 400) => {
-    if (err instanceof HttpError) sendJson({ error: err.message }, err.status);
-    else sendJson({ error: err instanceof Error ? err.message : String(err) }, fallback);
-  };
-
-  // Mutation endpoints only accept tiny config bodies ({label, group_name}
-  // etc). Cap at 64 KB to prevent memory/CPU abuse from an oversized payload.
-  const MAX_MUTATION_BODY = 64_000;
-  async function readJsonBody(): Promise<Record<string, unknown>> {
-    let body = "";
-    for await (const chunk of req) {
-      body += chunk;
-      if (body.length > MAX_MUTATION_BODY) {
-        // Destroy the socket so we stop reading further bytes from an
-        // oversized payload — unlike a plain throw, which lets the rest
-        // of the stream keep draining. Matches the /api/import/* pattern.
-        req.destroy();
-        throw new HttpError("Payload too large", 413);
-      }
-    }
-    if (!body) return {};
-    try { return JSON.parse(body) as Record<string, unknown>; }
-    catch { throw new HttpError("Invalid JSON body", 400); }
-  }
-
-  // Artifact endpoints (both reads and mutations) are localhost-only. A
-  // browser tab on some other site could otherwise fetch user data or
-  // trigger destructive actions via http://localhost:<port>/api/…. Mirrors
-  // the /mcp handler pattern: reject non-local origins outright; echo the
-  // origin back for local ones to override the wildcard CORS header set
-  // above.
-  const rejectIfNonLocalOrigin = (): boolean => {
-    const origin = req.headers.origin;
-    if (origin && !/^https?:\/\/(localhost|127\.0\.0\.1)(:\d+)?$/.test(origin)) {
-      sendJson({ error: "Forbidden origin" }, 403);
-      return true;
-    }
-    if (!origin) {
-      const remote = req.socket.remoteAddress || "";
-      const isLoopback =
-        remote === "127.0.0.1" ||
-        remote === "::1" ||
-        remote === "::ffff:127.0.0.1";
-      if (!isLoopback) {
-        sendJson({ error: "Forbidden — non-local origin" }, 403);
-        return true;
-      }
-    }
-    if (origin) res.setHeader("Access-Control-Allow-Origin", origin);
-    return false;
-  };
+  // /api/sessions/* — first extracted route bucket. Returns true when
+  // handled; falls through if no session route matched.
+  if (await tryHandleSessionRoute(req, res, url, ctx, {
+    sessionStore, spaceStore, artifactService, memoryProvider,
+  })) return;
 
   // GET /api/resolve-path?url=...  — resolve a serving URL to a filesystem path
   if (url.startsWith("/api/resolve-path")) {
@@ -930,279 +877,6 @@ async function handleHttpRequest(req: IncomingMessage, res: ServerResponse) {
         sendJson(memory, 201);
       } catch (err) {
         sendError(err);
-      }
-      return;
-    }
-  }
-
-  // GET /api/sessions — agent sessions captured by the watchers (#251).
-  // Read-only for 0.5.0; the home feed renders these. Local-origin only —
-  // session titles are derived from user prompts, which are private.
-  if (url === "/api/sessions" && req.method === "GET") {
-    if (rejectIfNonLocalOrigin()) return;
-    const rows = sessionStore.getAll();
-    // Join sources for sourceLabel — batched IN-list queries so the
-    // home feed can show "active project" tiles without a per-tile
-    // round trip. Sources are dedup'd because most sessions cluster
-    // around a small number of registered folders. Chunked at 500
-    // ids per batch to stay well below SQLite's 999-bound-variable
-    // ceiling on installs that haven't been recompiled with the
-    // higher 32_766 limit.
-    const sourceIds = [...new Set(rows.map((r) => r.source_id).filter((id): id is string => !!id))];
-    const SOURCE_BATCH = 500;
-    const sourceList = [];
-    for (let i = 0; i < sourceIds.length; i += SOURCE_BATCH) {
-      sourceList.push(...spaceStore.getSourcesByIds(sourceIds.slice(i, i + SOURCE_BATCH)));
-    }
-    const sourcesById = new Map(sourceList.map((s) => [s.id, s]));
-    sendJson(rows.map((row) => {
-      const src = row.source_id ? sourcesById.get(row.source_id) : null;
-      const label = src ? (src.label ?? (basename(src.path) || null)) : null;
-      return {
-        id: row.id,
-        spaceId: row.space_id,
-        sourceId: row.source_id ?? null,
-        sourceLabel: label,
-        cwd: row.cwd,
-        agent: row.agent,
-        title: row.title,
-        state: row.state,
-        startedAt: row.started_at,
-        endedAt: row.ended_at,
-        model: row.model,
-        lastEventAt: row.last_event_at,
-      };
-    }));
-    return;
-  }
-
-  // GET /api/sessions/search?q=…&session_id=…&limit=…
-  // R2 verbatim recall (#311). FTS5 over session_events.text. Mirrors the
-  // MCP `recall_transcripts` tool surface for the web UI.
-  // Local-origin only — transcripts are private user content.
-  {
-    const searchPath = url.split("?")[0];
-    if (searchPath === "/api/sessions/search" && req.method === "GET") {
-      if (rejectIfNonLocalOrigin()) return;
-      const parsed = new URL(req.url ?? "/", "http://localhost");
-      const q = parsed.searchParams.get("q") ?? "";
-      const scopeSession = parsed.searchParams.get("session_id") ?? undefined;
-      const limitRaw = parsed.searchParams.get("limit");
-      // Validate before clamping — Number("foo") is NaN, which Math.min/max
-      // propagate. Treat anything non-finite or non-positive as "use the
-      // store's default" rather than 400'ing.
-      let limit: number | undefined;
-      if (limitRaw !== null) {
-        const parsedLimit = Number(limitRaw);
-        if (Number.isFinite(parsedLimit) && parsedLimit >= 1) {
-          limit = Math.min(50, Math.floor(parsedLimit));
-        }
-      }
-      try {
-        const hits = sessionStore.searchEvents(q, { sessionId: scopeSession, limit });
-        // Rename `id` → `event_id` for the wire format (the web UI's
-        // ambient `id` is artefact id; explicit naming avoids confusion).
-        sendJson(hits.map((h) => ({
-          event_id: h.id,
-          session_id: h.session_id,
-          session_title: h.session_title,
-          role: h.role,
-          ts: h.ts,
-          snippet: h.snippet,
-        })));
-      } catch (err) {
-        sendError(err, 500);
-      }
-      return;
-    }
-  }
-
-  // GET /api/sessions/:id — single session row (or 404)
-  {
-    const m = url.match(/^\/api\/sessions\/([^/]+)$/);
-    if (m && req.method === "GET") {
-      if (rejectIfNonLocalOrigin()) return;
-      const row = sessionStore.getById(m[1]);
-      if (!row) {
-        res.writeHead(404, { "Content-Type": "application/json" });
-        res.end(JSON.stringify({ error: "session not found" }));
-        return;
-      }
-      const src = row.source_id ? spaceStore.getSourceById(row.source_id) : undefined;
-      const sourceLabel = src ? (src.label ?? (basename(src.path) || null)) : null;
-      sendJson({
-        id: row.id,
-        spaceId: row.space_id,
-        sourceId: row.source_id ?? null,
-        sourceLabel,
-        cwd: row.cwd,
-        agent: row.agent,
-        title: row.title,
-        state: row.state,
-        startedAt: row.started_at,
-        endedAt: row.ended_at,
-        model: row.model,
-        lastEventAt: row.last_event_at,
-      });
-      return;
-    }
-  }
-
-  // GET /api/sessions/:id/events — transcript events (oldest first within
-  // the returned slice). The `raw` JSONL line is dropped because long
-  // sessions can ship 50+MB of raw blobs; clients lazy-fetch raw via
-  // /events/:eventId when they expand a tool turn.
-  //
-  // Cursors:
-  //   ?before=<id> — events with id < before, latest N (load older on scroll up)
-  //   ?after=<id>  — events with id > after, oldest N (live append)
-  //   neither     — latest N (bootstrap)
-  // ?limit=N defaults to 1000.
-  {
-    // Strip the query string before path matching — the `$` anchor in the
-    // regex would otherwise reject any URL with `?...`. Pre-existing bug:
-    // `?limit=N` was always silently ignored before this fix.
-    const eventsPath = url.split("?")[0];
-    const m = eventsPath.match(/^\/api\/sessions\/([^/]+)\/events$/);
-    if (m && req.method === "GET") {
-      if (rejectIfNonLocalOrigin()) return;
-      const parsed = new URL(req.url ?? "/", "http://localhost");
-      const limitParam = parsed.searchParams.get("limit");
-      const limit = limitParam && Number.isFinite(Number(limitParam))
-        ? Math.max(1, Math.min(10_000, Number(limitParam)))
-        : 1000;
-      const beforeParam = parsed.searchParams.get("before");
-      const afterParam = parsed.searchParams.get("after");
-      const aroundParam = parsed.searchParams.get("around");
-      const before = beforeParam && Number.isFinite(Number(beforeParam))
-        ? Number(beforeParam) : null;
-      const after = afterParam && Number.isFinite(Number(afterParam))
-        ? Number(afterParam) : null;
-      const around = aroundParam && Number.isFinite(Number(aroundParam))
-        ? Number(aroundParam) : null;
-      let events;
-      if (around !== null) {
-        // Centred window: split the budget so the merged result is at
-        // most `limit` events. The target is included in the "older"
-        // half (id <= around), so olderLimit gets the ceil. Sort ASC
-        // by id so the transcript renders chronologically regardless
-        // of how the underlying statements ordered their result sets.
-        const olderLimit = Math.max(1, Math.ceil(limit / 2));
-        const newerLimit = Math.max(0, limit - olderLimit);
-        const older = sessionStore.getEventsBeforeBySession(m[1], around + 1, olderLimit);
-        const newer = newerLimit > 0
-          ? sessionStore.getEventsAfterBySession(m[1], around, newerLimit)
-          : [];
-        events = [...older, ...newer].sort((a, b) => a.id - b.id);
-      } else if (before !== null) {
-        events = sessionStore.getEventsBeforeBySession(m[1], before, limit);
-      } else if (after !== null) {
-        events = sessionStore.getEventsAfterBySession(m[1], after, limit);
-      } else {
-        events = sessionStore.getEventsBySession(m[1], { limit });
-      }
-      sendJson(events.map((e) => ({
-        id: e.id,
-        sessionId: e.session_id,
-        role: e.role,
-        text: e.text,
-        ts: e.ts,
-        raw: null as string | null,
-      })));
-      return;
-    }
-  }
-
-  // GET /api/sessions/:id/events/:eventId — single event WITH raw JSONL.
-  // Exists so the inspector can lazily load the raw blob for tool-call
-  // expand without paying for it on every transcript fetch.
-  {
-    const m = url.match(/^\/api\/sessions\/([^/]+)\/events\/(\d+)$/);
-    if (m && req.method === "GET") {
-      if (rejectIfNonLocalOrigin()) return;
-      const eventId = Number(m[2]);
-      const ev = sessionStore.getEventById(m[1], eventId);
-      if (!ev) {
-        res.writeHead(404, { "Content-Type": "application/json" });
-        res.end(JSON.stringify({ error: "event not found" }));
-        return;
-      }
-      sendJson({
-        id: ev.id,
-        sessionId: ev.session_id,
-        role: ev.role,
-        text: ev.text,
-        ts: ev.ts,
-        raw: ev.raw,
-      });
-      return;
-    }
-  }
-
-  // GET /api/sessions/:id/artifacts — touched artefacts joined with artifact metadata
-  {
-    const m = url.match(/^\/api\/sessions\/([^/]+)\/artifacts$/);
-    if (m && req.method === "GET") {
-      if (rejectIfNonLocalOrigin()) return;
-      const touches = sessionStore.getArtifactsBySession(m[1]);
-      const uniqueIds = Array.from(new Set(touches.map((t) => t.artifact_id)));
-      const artifacts = await artifactService.getArtifactsByIds(uniqueIds);
-      const byId = new Map(artifacts.map((a) => [a.id, a]));
-      sendJson(touches.flatMap((t) => {
-        const a = byId.get(t.artifact_id);
-        if (!a) return [];
-        return [{
-          id: t.id,
-          sessionId: t.session_id,
-          artifactId: t.artifact_id,
-          role: t.role,
-          whenAt: t.when_at,
-          artifact: a,
-        }];
-      }));
-      return;
-    }
-  }
-
-  // GET /api/sessions/:id/memory — memories associated with this session.
-  // R6 traceable recall (#310): returns {written, pulled} where written
-  // are memories whose source_session_id == :id and pulled are memories
-  // this session retrieved via recall(). Each memory has its source
-  // session title resolved across the memory↔sessions DB boundary so
-  // the UI can render "from <title>" without a second round trip.
-  {
-    const m = url.match(/^\/api\/sessions\/([^/]+)\/memory$/);
-    if (m && req.method === "GET") {
-      if (rejectIfNonLocalOrigin()) return;
-      try {
-        const [written, pulled] = await Promise.all([
-          memoryProvider.getBySourceSession(m[1]),
-          memoryProvider.getRecalledBySession(m[1]),
-        ]);
-        // Resolve source_session_title for every memory we're about to
-        // return. Batched: collect distinct session ids, fetch titles
-        // once, then attach. memory.db and oyster.db are separate, so
-        // this stitch happens at the API layer.
-        const sourceIds = new Set<string>();
-        for (const memory of [...written, ...pulled]) {
-          if (memory.source_session_id) sourceIds.add(memory.source_session_id);
-        }
-        const titleById = new Map<string, string | null>();
-        for (const sid of sourceIds) {
-          titleById.set(sid, sessionStore.getById(sid)?.title ?? null);
-        }
-        const enrich = (memory: typeof written[number]) => ({
-          ...memory,
-          source_session_title: memory.source_session_id
-            ? (titleById.get(memory.source_session_id) ?? null)
-            : null,
-        });
-        sendJson({
-          written: written.map(enrich),
-          pulled: pulled.map(enrich),
-        });
-      } catch (err) {
-        sendError(err, 500);
       }
       return;
     }
